Moved Unit and Quantity into _unit.py. Added methods Expression.is_name() and .is_number().
