+++ myokit/__init__.py
@@ -388,7 +388,7 @@ from ._model_api import ( # noqa
     Variable,
 )
 
-# Expressions and units
+# Expressions
 from ._expressions import (  # noqa
     Abs,
     ACos,
@@ -429,14 +429,17 @@ from ._expressions import (  # noqa
     PrefixExpression,
     PrefixMinus,
     PrefixPlus,
-    Quantity,
     Quotient,
     Remainder,
     Sin,
     Sqrt,
     Tan,
-    Unit,
+)
 
+# Unit and quantity
+from ._unit import (  # noqa
+    Quantity,
+    Unit,
 )
 
 # Pacing protocol
