+++ myokit/tests/test_expressions.py
@@ -24,9 +24,7 @@ except AttributeError:
 
 
 class ExpressionTest(unittest.TestCase):
-    """
-    Tests various methods of the :class:`myokit.Expression` class.
-    """
+    """ Tests various methods of the :class:`myokit.Expression` class. """
 
     def test_contains_type(self):
         # Test :meth:`Expression.contains_type`.
@@ -287,9 +285,7 @@ class ExpressionTest(unittest.TestCase):
 
 
 class NumberTest(unittest.TestCase):
-    """
-    Tests myokit.Number.
-    """
+    """ Tests myokit.Number. """
 
     def test_basic(self):
         # Test construction, other basics.
@@ -411,6 +407,23 @@ class NumberTest(unittest.TestCase):
             x.eval_unit(myokit.UNIT_STRICT), myokit.units.dimensionless)
         self.assertEqual(y.eval_unit(myokit.UNIT_STRICT), myokit.units.ampere)
 
+    def test_is_name_is_number(self):
+        # Test is_name() and is_number
+
+        x = myokit.Number(3, myokit.units.volt)
+        self.assertFalse(x.is_name())
+        v = myokit.Model().add_component('c').add_variable('x')
+        self.assertFalse(x.is_name(v))
+
+        self.assertTrue(x.is_number())
+        self.assertFalse(x.is_number(12))
+        self.assertFalse(x.is_number(0))
+        self.assertTrue(x.is_number(3))
+
+        x = myokit.Number(0)
+        self.assertTrue(x.is_number())
+        self.assertTrue(x.is_number(0))
+
     def test_tree_str(self):
         # Test Number.tree_str()
 
@@ -424,9 +437,7 @@ class NumberTest(unittest.TestCase):
 
 
 class NameTest(unittest.TestCase):
-    """
-    Tests myokit.Name.
-    """
+    """ Tests myokit.Name. """
 
     def test_basics(self):
         # Test creation, representation,
@@ -588,6 +599,19 @@ class NameTest(unittest.TestCase):
         self.assertEqual(
             y.rhs().eval_unit(myokit.UNIT_STRICT), myokit.units.dimensionless)
 
+    def test_is_name_is_number(self):
+        # Test is_name() and is_number
+
+        v = myokit.Model().add_component('c').add_variable('v')
+        x = myokit.Name(v)
+        self.assertTrue(x.is_name())
+        self.assertTrue(x.is_name(v))
+        w = v.parent().add_variable('w')
+        self.assertFalse(x.is_name(w))
+
+        self.assertFalse(x.is_number())
+        self.assertFalse(x.is_number(0))
+
     def test_rhs(self):
         # Test Name.rhs().
 
@@ -623,9 +647,8 @@ class NameTest(unittest.TestCase):
 
 
 class DerivativeTest(unittest.TestCase):
-    """
-    Tests myokit.Derivative.
-    """
+    """ Tests myokit.Derivative. """
+
     def test_basic(self):
         # Test creation.
 
@@ -717,6 +740,20 @@ class DerivativeTest(unittest.TestCase):
         t.set_unit(None)
         self.assertEqual(d.eval_unit(), myokit.units.volt)
         self.assertEqual(d.eval_unit(s), myokit.units.volt)
+
+    def test_is_name_is_number(self):
+        # Test is_name() and is_number
+
+        v = myokit.Model().add_component('c').add_variable('v')
+        x = myokit.Derivative(myokit.Name(v))
+        self.assertFalse(x.is_name())
+        self.assertFalse(x.is_name(v))
+        w = v.parent().add_variable('w')
+        self.assertFalse(x.is_name(w))
+
+        self.assertFalse(x.is_number())
+        self.assertFalse(x.is_number(0))
+        self.assertFalse(x.is_number(1))
 
     def test_rhs(self):
         # Test Derivative.rhs()
@@ -814,6 +851,18 @@ class PrefixPlusTest(unittest.TestCase):
         self.assertEqual(
             y.rhs().eval_unit(myokit.UNIT_STRICT), myokit.units.Newton)
 
+    def test_is_name_is_number(self):
+        # Test is_name() and is_number()
+
+        x = myokit.PrefixPlus(myokit.Number(1))
+        self.assertFalse(x.is_name())
+        v = myokit.Model().add_component('c').add_variable('v')
+        self.assertFalse(x.is_name(v))
+
+        self.assertFalse(x.is_number())
+        self.assertFalse(x.is_number(0))
+        self.assertFalse(x.is_number(1))
+
     def test_tree_str(self):
         # Test PrefixPlus.tree_str()
 
@@ -833,7 +882,7 @@ class PrefixMinusTest(unittest.TestCase):
     Tests myokit.PrefixMinus.
     """
     def test_clone(self):
-        """ Test PrefixMinus.clone(). """
+        # Test PrefixMinus.clone().
         x = myokit.PrefixMinus(myokit.Number(3))
         y = x.clone()
         self.assertIsNot(y, x)
@@ -855,7 +904,8 @@ class PrefixMinusTest(unittest.TestCase):
         self.assertEqual(y, myokit.PrefixMinus(j))
 
     def test_bracket(self):
-        """ Test PrefixMinus.bracket(). """
+        # Test PrefixMinus.bracket().
+
         i = myokit.Number(1)
         x = myokit.PrefixMinus(i)
         self.assertFalse(x.bracket(i))
@@ -865,16 +915,14 @@ class PrefixMinusTest(unittest.TestCase):
         self.assertRaises(ValueError, x.bracket, myokit.Number(1))
 
     def test_eval(self):
-        """
-        Test PrefixMinus evaluation.
-        """
+        # Test PrefixMinus evaluation.
+
         x = myokit.PrefixMinus(myokit.Number(2))
         self.assertEqual(x.eval(), -2)
 
     def test_eval_unit(self):
-        """
-        Test PrefixMinus.eval_unit().
-        """
+        # Test PrefixMinus.eval_unit().
+
         # Mini model
         m = myokit.Model()
         c = m.add_component('c')
@@ -893,8 +941,20 @@ class PrefixMinusTest(unittest.TestCase):
         self.assertEqual(
             y.rhs().eval_unit(myokit.UNIT_STRICT), myokit.units.Newton)
 
+    def test_is_name_is_number(self):
+        # Test is_name() and is_number()
+
+        x = myokit.PrefixMinus(myokit.Number(1))
+        self.assertFalse(x.is_name())
+        v = myokit.Model().add_component('c').add_variable('v')
+        self.assertFalse(x.is_name(v))
+
+        self.assertFalse(x.is_number())
+        self.assertFalse(x.is_number(0))
+        self.assertFalse(x.is_number(1))
+
     def test_tree_str(self):
-        """ Test PrefixMinus.tree_str() """
+        # Test PrefixMinus.tree_str()
         # Test simple
         x = myokit.PrefixMinus(myokit.Number(1))
         self.assertEqual(x.tree_str(), '-\n  1\n')
@@ -911,7 +971,7 @@ class PlusTest(unittest.TestCase):
     Tests myokit.Plus.
     """
     def test_clone(self):
-        """ Test Plus.clone(). """
+        # Test Plus.clone().
         i = myokit.Number(3)
         j = myokit.Number(4)
         x = myokit.Plus(i, j)
@@ -936,7 +996,7 @@ class PlusTest(unittest.TestCase):
         self.assertEqual(y, myokit.Plus(i, i))
 
     def test_bracket(self):
-        """ Test Plus.bracket(). """
+        # Test Plus.bracket().
         i = myokit.Number(1)
         j = myokit.parse_expression('1 + 2')
         x = myokit.Plus(i, j)
@@ -945,16 +1005,14 @@ class PlusTest(unittest.TestCase):
         self.assertRaises(ValueError, x.bracket, myokit.Number(3))
 
     def test_eval(self):
-        """
-        Test Plus evaluation.
-        """
+        # Test Plus evaluation.
         x = myokit.Plus(myokit.Number(1), myokit.Number(2))
         self.assertEqual(x.eval(), 3)
         x = myokit.Plus(myokit.Number(1), myokit.PrefixMinus(myokit.Number(2)))
         self.assertEqual(x.eval(), -1)
 
     def test_eval_unit(self):
-        """ Test Plus.eval_unit(). """
+        # Test Plus.eval_unit().
         # Create mini model
         m = myokit.Model()
         c = m.add_component('c')
@@ -1011,8 +1069,20 @@ class PlusTest(unittest.TestCase):
             self.assertEqual(token[2], 3)
             self.assertEqual(token[3], 11)
 
+    def test_is_name_is_number(self):
+        # Test is_name() and is_number()
+
+        x = myokit.Plus(myokit.Number(1), myokit.Number(0))
+        self.assertFalse(x.is_name())
+        v = myokit.Model().add_component('c').add_variable('v')
+        self.assertFalse(x.is_name(v))
+
+        self.assertFalse(x.is_number())
+        self.assertFalse(x.is_number(0))
+        self.assertFalse(x.is_number(1))
+
     def test_tree_str(self):
-        """ Test Plus.tree_str(). """
+        # Test Plus.tree_str().
         # Test simple
         x = myokit.Plus(myokit.Number(1), myokit.Number(2))
         self.assertEqual(x.tree_str(), '+\n  1\n  2\n')
@@ -1029,14 +1099,14 @@ class MinusTest(unittest.TestCase):
     Tests myokit.Minus.
     """
     def test_eval(self):
-        """
-        Test Minus evaluation.
-        """
+        # Test Minus evaluation.
+
         x = myokit.Minus(myokit.Number(1), myokit.Number(2))
         self.assertEqual(x.eval(), -1)
 
     def test_eval_unit(self):
-        """ Test Minus.eval_unit(). """
+        # Test Minus.eval_unit().
+
         # Create mini model
         m = myokit.Model()
         c = m.add_component('c')
@@ -1076,8 +1146,20 @@ class MinusTest(unittest.TestCase):
         self.assertEqual(
             z.rhs().eval_unit(myokit.UNIT_STRICT), myokit.units.dimensionless)
 
+    def test_is_name_is_number(self):
+        # Test is_name() and is_number()
+
+        x = myokit.Minus(myokit.Number(2), myokit.Number(3))
+        self.assertFalse(x.is_name())
+        v = myokit.Model().add_component('c').add_variable('v')
+        self.assertFalse(x.is_name(v))
+
+        self.assertFalse(x.is_number())
+        self.assertFalse(x.is_number(0))
+        self.assertFalse(x.is_number(1))
+
     def test_tree_str(self):
-        """ Test Minus.tree_str(). """
+        # Test Minus.tree_str().
         # Test simple
         x = myokit.Minus(myokit.Number(1), myokit.Number(2))
         self.assertEqual(x.tree_str(), '-\n  1\n  2\n')
@@ -1094,14 +1176,14 @@ class MultiplyTest(unittest.TestCase):
     Tests myokit.Multiply.
     """
     def test_eval(self):
-        """
-        Test Multiply evaluation.
-        """
+        # Test Multiply evaluation.
+
         x = myokit.Multiply(myokit.Number(1), myokit.Number(2))
         self.assertEqual(x.eval(), 2)
 
     def test_eval_unit(self):
-        """ Test Multiply.eval_unit(). """
+        # Test Multiply.eval_unit().
+
         # Create mini model
         m = myokit.Model()
         c = m.add_component('c')
@@ -1143,7 +1225,8 @@ class MultiplyTest(unittest.TestCase):
             z.rhs().eval_unit(myokit.UNIT_STRICT), myokit.units.dimensionless)
 
     def test_tree_str(self):
-        """ Test Multiply.tree_str(). """
+        # Test Multiply.tree_str().
+
         # Test simple
         x = myokit.Multiply(myokit.Number(1), myokit.Number(2))
         self.assertEqual(x.tree_str(), '*\n  1\n  2\n')
@@ -1160,7 +1243,7 @@ class DivideTest(unittest.TestCase):
     Tests myokit.Divide.
     """
     def test_tree_str(self):
-        """ Test Divide.tree_str(). """
+        # Test Divide.tree_str().
         # Test simple
         x = myokit.Divide(myokit.Number(1), myokit.Number(2))
         self.assertEqual(x.tree_str(), '/\n  1\n  2\n')
@@ -1172,14 +1255,14 @@ class DivideTest(unittest.TestCase):
         self.assertEqual(x.tree_str(), '/\n  1\n  /\n    2\n    3\n')
 
     def test_eval(self):
-        """
-        Test Divide evaluation.
-        """
+        # Test Divide evaluation.
+
         x = myokit.Divide(myokit.Number(1), myokit.Number(2))
         self.assertEqual(x.eval(), 0.5)
 
     def test_eval_unit(self):
-        """ Test Divide.eval_unit(). """
+        # Test Divide.eval_unit().
+
         # Create mini model
         m = myokit.Model()
         c = m.add_component('c')
@@ -1226,14 +1309,13 @@ class QuotientTest(unittest.TestCase):
     Tests myokit.Quotient.
     """
     def test_eval(self):
-        """
-        Test Quotient evaluation.
-        """
+        # Test Quotient evaluation.
         x = myokit.Quotient(myokit.Number(7), myokit.Number(2))
         self.assertEqual(x.eval(), 3.0)
 
     def test_eval_unit(self):
-        """ Test Quotient.eval_unit(). """
+        # Test Quotient.eval_unit().
+
         # Create mini model
         m = myokit.Model()
         c = m.add_component('c')
@@ -1275,7 +1357,7 @@ class QuotientTest(unittest.TestCase):
             z.rhs().eval_unit(myokit.UNIT_STRICT), myokit.units.dimensionless)
 
     def test_tree_str(self):
-        """ Test Quotient.tree_str(). """
+        # Test Quotient.tree_str().
         # Test simple
         x = myokit.Quotient(myokit.Number(1), myokit.Number(2))
         self.assertEqual(x.tree_str(), '//\n  1\n  2\n')
@@ -1292,14 +1374,14 @@ class RemainderTest(unittest.TestCase):
     Tests myokit.Remainder.
     """
     def test_eval(self):
-        """
-        Test Divide evaluation.
-        """
+        # Test Divide evaluation.
+
         x = myokit.Remainder(myokit.Number(7), myokit.Number(4))
         self.assertEqual(x.eval(), 3.0)
 
     def test_eval_unit(self):
-        """ Test Remainder.eval_unit(). """
+        # Test Remainder.eval_unit().
+
         # Create mini model
         m = myokit.Model()
         c = m.add_component('c')
@@ -1341,7 +1423,8 @@ class RemainderTest(unittest.TestCase):
             z.rhs().eval_unit(myokit.UNIT_STRICT), myokit.units.dimensionless)
 
     def test_tree_str(self):
-        """ Test Remainder.tree_str(). """
+        # Test Remainder.tree_str().
+
         # Test simple
         x = myokit.Remainder(myokit.Number(1), myokit.Number(2))
         self.assertEqual(x.tree_str(), '%\n  1\n  2\n')
@@ -1358,7 +1441,7 @@ class PowerTest(unittest.TestCase):
     Tests myokit.Power.
     """
     def test_clone(self):
-        """ Test Power.clone(). """
+        # Test Power.clone().
         i = myokit.Number(3)
         j = myokit.Number(4)
         x = myokit.Power(i, j)
@@ -1383,7 +1466,7 @@ class PowerTest(unittest.TestCase):
         self.assertEqual(y, myokit.Power(i, i))
 
     def test_bracket(self):
-        """ Test Power.bracket(). """
+        # Test Power.bracket().
         i = myokit.Number(1)
         j = myokit.parse_expression('1 + 2')
         x = myokit.Power(i, j)
@@ -1392,9 +1475,8 @@ class PowerTest(unittest.TestCase):
         self.assertRaises(ValueError, x.bracket, myokit.Number(3))
 
     def test_eval(self):
-        """
-        Test Power evaluation.
-        """
+        # Test Power evaluation.
+
         x = myokit.Power(myokit.Number(2), myokit.Number(3))
         self.assertEqual(x.eval(), 8)
         x = myokit.Power(
@@ -1402,7 +1484,8 @@ class PowerTest(unittest.TestCase):
         self.assertEqual(x.eval(), 0.5)
 
     def test_eval_unit(self):
-        """ Test Power.eval_unit(). """
+        # Test Power.eval_unit().
+
         # Create mini model
         m = myokit.Model()
         c = m.add_component('c')
@@ -1441,8 +1524,21 @@ class PowerTest(unittest.TestCase):
         self.assertEqual(
             z.rhs().eval_unit(myokit.UNIT_STRICT), myokit.units.dimensionless)
 
+    def test_is_name_is_number(self):
+        # Test is_name() and is_number()
+
+        x = myokit.Power(myokit.Number(2), myokit.Number(3))
+        self.assertFalse(x.is_name())
+        v = myokit.Model().add_component('c').add_variable('v')
+        self.assertFalse(x.is_name(v))
+
+        self.assertFalse(x.is_number())
+        self.assertFalse(x.is_number(0))
+        self.assertFalse(x.is_number(1))
+
     def test_tree_str(self):
-        """ Test Power.tree_str(). """
+        # Test Power.tree_str().
+
         # Test simple
         x = myokit.Power(myokit.Number(1), myokit.Number(2))
         self.assertEqual(x.tree_str(), '^\n  1\n  2\n')
@@ -1459,14 +1555,14 @@ class SqrtTest(unittest.TestCase):
     Tests myokit.Sqrt.
     """
     def test_creation(self):
-        """ Test Sqrt creation. """
+        # Test Sqrt creation.
         myokit.Sqrt(myokit.Number(1))
         self.assertRaisesRegex(
             myokit.IntegrityError, 'wrong number', myokit.Sqrt,
             myokit.Number(1), myokit.Number(2))
 
     def test_clone(self):
-        """ Test Sqrt.clone(). """
+        # Test Sqrt.clone().
         i = myokit.Number(3)
         j = myokit.Number(10)
         x = myokit.Sqrt(i)
@@ -1487,7 +1583,7 @@ class SqrtTest(unittest.TestCase):
         self.assertEqual(y, z)
 
     def test_bracket(self):
-        """ Test Sqrt.bracket(). """
+        # Test Sqrt.bracket().
         i = myokit.Number(1)
         j = myokit.parse_expression('1 + 2')
         x = myokit.Sqrt(i)
@@ -1497,14 +1593,14 @@ class SqrtTest(unittest.TestCase):
         self.assertRaises(ValueError, x.bracket, myokit.Number(3))
 
     def test_eval(self):
-        """
-        Test Sqrt evaluation.
-        """
+        # Test Sqrt evaluation.
+
         x = myokit.Sqrt(myokit.Number(9))
         self.assertEqual(x.eval(), 3)
 
     def test_eval_unit(self):
-        """ Test Sqrt.eval_unit(). """
+        # Test Sqrt.eval_unit().
+
         # Create mini model
         m = myokit.Model()
         c = m.add_component('c')
@@ -1533,7 +1629,7 @@ class SqrtTest(unittest.TestCase):
         self.assertEqual(z.rhs().eval_unit(s), myokit.units.dimensionless)
 
     def test_tree_str(self):
-        """ Test Sqrt.tree_str(). """
+        # Test Sqrt.tree_str().
         # Test simple
         x = myokit.Sqrt(myokit.Number(2))
         self.assertEqual(x.tree_str(), 'sqrt\n  2\n')
@@ -1550,14 +1646,14 @@ class ExpTest(unittest.TestCase):
     Tests myokit.Exp.
     """
     def test_creation(self):
-        """ Test Exp creation. """
+        # Test Exp creation.
         myokit.Exp(myokit.Number(1))
         self.assertRaisesRegex(
             myokit.IntegrityError, 'wrong number', myokit.Exp,
             myokit.Number(1), myokit.Number(2))
 
     def test_clone(self):
-        """ Test Exp.clone(). """
+        # Test Exp.clone().
         i = myokit.Number(3)
         j = myokit.Number(10)
         x = myokit.Exp(i)
@@ -1578,7 +1674,7 @@ class ExpTest(unittest.TestCase):
         self.assertEqual(y, z)
 
     def test_bracket(self):
-        """ Test Exp.bracket(). """
+        # Test Exp.bracket().
         i = myokit.Number(1)
         j = myokit.parse_expression('1 + 2')
         x = myokit.Exp(i)
@@ -1588,12 +1684,12 @@ class ExpTest(unittest.TestCase):
         self.assertRaises(ValueError, x.bracket, myokit.Number(3))
 
     def test_eval(self):
-        """ Test Exp.eval(). """
+        # Test Exp.eval().
         x = myokit.Exp(myokit.Number(9))
         self.assertEqual(x.eval(), np.exp(9))
 
     def test_eval_unit(self):
-        """ Test Exp.eval_unit(). """
+        # Test Exp.eval_unit().
         # Create mini model
         m = myokit.Model()
         c = m.add_component('c')
@@ -1619,8 +1715,20 @@ class ExpTest(unittest.TestCase):
         x.set_unit(None)
         self.assertEqual(z.rhs().eval_unit(s), myokit.units.dimensionless)
 
+    def test_is_name_is_number(self):
+        # Test is_name() and is_number()
+
+        x = myokit.Exp(myokit.Number(7))
+        self.assertFalse(x.is_name())
+        v = myokit.Model().add_component('c').add_variable('v')
+        self.assertFalse(x.is_name(v))
+
+        self.assertFalse(x.is_number())
+        self.assertFalse(x.is_number(0))
+        self.assertFalse(x.is_number(1))
+
     def test_tree_str(self):
-        """ Test Exp.tree_str(). """
+        # Test Exp.tree_str().
         # Test simple
         x = myokit.Exp(myokit.Number(2))
         self.assertEqual(x.tree_str(), 'exp\n  2\n')
@@ -1637,7 +1745,7 @@ class LogTest(unittest.TestCase):
     Tests myokit.Log.
     """
     def test_creation(self):
-        """ Test Log creation. """
+        # Test Log creation.
         myokit.Log(myokit.Number(1))
         myokit.Log(myokit.Number(1), myokit.Number(2))
         self.assertRaisesRegex(
@@ -1645,7 +1753,7 @@ class LogTest(unittest.TestCase):
             myokit.Number(1), myokit.Number(2), myokit.Number(3))
 
     def test_clone(self):
-        """ Test Log.clone(). """
+        # Test Log.clone().
         # Test with one operand
         i = myokit.Number(3)
         j = myokit.Number(10)
@@ -1689,7 +1797,7 @@ class LogTest(unittest.TestCase):
         self.assertEqual(y, myokit.Log(i, i))
 
     def test_bracket(self):
-        """ Test Log.bracket(). """
+        # Test Log.bracket().
         i = myokit.Number(1)
         j = myokit.parse_expression('1 + 2')
 
@@ -1707,7 +1815,7 @@ class LogTest(unittest.TestCase):
         self.assertRaises(ValueError, x.bracket, myokit.Number(3))
 
     def test_eval(self):
-        """ Test Exp.eval(). """
+        # Test Exp.eval().
         # One argument
         x = myokit.Log(myokit.Number(9))
         self.assertEqual(x.eval(), np.log(9))
@@ -1717,7 +1825,7 @@ class LogTest(unittest.TestCase):
         self.assertEqual(x.eval(), np.log(9) / np.log(5))
 
     def test_eval_unit(self):
-        """ Test Log.eval_unit(). """
+        # Test Log.eval_unit().
         # Create mini model
         m = myokit.Model()
         c = m.add_component('c')
@@ -1780,7 +1888,7 @@ class LogTest(unittest.TestCase):
         self.assertEqual(z.rhs().eval_unit(s), myokit.units.dimensionless)
 
     def test_tree_str(self):
-        """ Test Log.tree_str(). """
+        # Test Log.tree_str().
         # Test simple
         x = myokit.Log(myokit.Number(2))
         self.assertEqual(x.tree_str(), 'log\n  2\n')
@@ -1797,12 +1905,12 @@ class Log10Test(unittest.TestCase):
     Tests myokit.Log10.
     """
     def test_eval(self):
-        """ Test Log10.eval(). """
+        # Test Log10.eval().
         x = myokit.Log10(myokit.Number(9))
         self.assertEqual(x.eval(), np.log10(9))
 
     def test_tree_str(self):
-        """ Test Log10.tree_str(). """
+        # Test Log10.tree_str().
         x = myokit.Log10(myokit.Number(2))
         self.assertEqual(x.tree_str(), 'log10\n  2\n')
 
@@ -1812,12 +1920,12 @@ class SinTest(unittest.TestCase):
     Tests myokit.Sin.
     """
     def test_eval(self):
-        """ Test Sin.eval(). """
+        # Test Sin.eval().
         x = myokit.Sin(myokit.Number(9))
         self.assertEqual(x.eval(), np.sin(9))
 
     def test_tree_str(self):
-        """ Test Sin.tree_str(). """
+        # Test Sin.tree_str().
         x = myokit.Sin(myokit.Number(2))
         self.assertEqual(x.tree_str(), 'sin\n  2\n')
 
@@ -1827,89 +1935,79 @@ class CosTest(unittest.TestCase):
     Tests myokit.Cos.
     """
     def test_eval(self):
-        """ Test Cos.eval(). """
+        # Test Cos.eval().
         x = myokit.Cos(myokit.Number(9))
         self.assertEqual(x.eval(), np.cos(9))
 
     def test_tree_str(self):
-        """ Test Cos.tree_str(). """
+        # Test Cos.tree_str().
         x = myokit.Cos(myokit.Number(2))
         self.assertEqual(x.tree_str(), 'cos\n  2\n')
 
 
 class TanTest(unittest.TestCase):
-    """
-    Tests myokit.Tan.
-    """
+    """ Tests myokit.Tan. """
     def test_eval(self):
-        """ Test Tan.eval(). """
+        # Test Tan.eval().
         x = myokit.Tan(myokit.Number(9))
         self.assertEqual(x.eval(), np.tan(9))
 
     def test_tree_str(self):
-        """ Test Tan.tree_str(). """
+        # Test Tan.tree_str().
         x = myokit.Tan(myokit.Number(2))
         self.assertEqual(x.tree_str(), 'tan\n  2\n')
 
 
 class ASinTest(unittest.TestCase):
-    """
-    Tests myokit.ASin.
-    """
+    """ Tests myokit.ASin. """
     def test_eval(self):
-        """ Test ASin.eval(). """
+        # Test ASin.eval().
         x = myokit.ASin(myokit.Number(0.9))
         self.assertEqual(x.eval(), np.arcsin(0.9))
 
     def test_tree_str(self):
-        """ Test ASin.tree_str(). """
+        # Test ASin.tree_str().
         x = myokit.ASin(myokit.Number(0.5))
         self.assertEqual(x.tree_str(), 'asin\n  0.5\n')
 
 
 class ACosTest(unittest.TestCase):
-    """
-    Tests myokit.ACos.
-    """
+    """ Tests myokit.ACos. """
     def test_eval(self):
-        """ Test ACos.eval(). """
+        # Test ACos.eval().
         x = myokit.ACos(myokit.Number(0.9))
         self.assertEqual(x.eval(), np.arccos(0.9))
 
     def test_tree_str(self):
-        """ Test ACos.tree_str(). """
+        # Test ACos.tree_str().
         x = myokit.ACos(myokit.Number(0.5))
         self.assertEqual(x.tree_str(), 'acos\n  0.5\n')
 
 
 class ATanTest(unittest.TestCase):
-    """
-    Tests myokit.ATan.
-    """
+    """ Tests myokit.ATan. """
     def test_eval(self):
-        """ Test ATan.eval(). """
+        # Test ATan.eval().
         x = myokit.ATan(myokit.Number(0.9))
         self.assertEqual(x.eval(), np.arctan(0.9))
 
     def test_tree_str(self):
-        """ Test ATan.tree_str(). """
+        # Test ATan.tree_str().
         x = myokit.ATan(myokit.Number(0.5))
         self.assertEqual(x.tree_str(), 'atan\n  0.5\n')
 
 
 class FloorTest(unittest.TestCase):
-    """
-    Tests myokit.Floor.
-    """
+    """ Tests myokit.Floor. """
+
     def test_eval(self):
-        """ Test Floor.eval(). """
+        # Test Floor.eval().
         x = myokit.Floor(myokit.Number(0.9))
         self.assertEqual(x.eval(), 0)
 
     def test_eval_unit(self):
-        """
-        Test Floor.eval_unit().
-        """
+        # Test Floor.eval_unit().
+
         # Mini model
         m = myokit.Model()
         c = m.add_component('c')
@@ -1929,24 +2027,22 @@ class FloorTest(unittest.TestCase):
             y.rhs().eval_unit(myokit.UNIT_STRICT), myokit.units.Newton)
 
     def test_tree_str(self):
-        """ Test Floor.tree_str(). """
+        # Test Floor.tree_str().
         x = myokit.Floor(myokit.Number(0.5))
         self.assertEqual(x.tree_str(), 'floor\n  0.5\n')
 
 
 class CeilTest(unittest.TestCase):
-    """
-    Tests myokit.Ceil.
-    """
+    """ Tests myokit.Ceil. """
+
     def test_eval(self):
-        """ Test Ceil.eval(). """
+        # Test Ceil.eval().
         x = myokit.Ceil(myokit.Number(0.9))
         self.assertEqual(x.eval(), 1)
 
     def test_eval_unit(self):
-        """
-        Test Ceil.eval_unit().
-        """
+        # Test Ceil.eval_unit().
+
         # Mini model
         m = myokit.Model()
         c = m.add_component('c')
@@ -1966,26 +2062,23 @@ class CeilTest(unittest.TestCase):
             y.rhs().eval_unit(myokit.UNIT_STRICT), myokit.units.Newton)
 
     def test_tree_str(self):
-        """ Test Ceil.tree_str(). """
+        # Test Ceil.tree_str().
         x = myokit.Ceil(myokit.Number(0.5))
         self.assertEqual(x.tree_str(), 'ceil\n  0.5\n')
 
 
 class AbsTest(unittest.TestCase):
-    """
-    Tests myokit.Abs.
-    """
+    """ Tests myokit.Abs. """
     def test_eval(self):
-        """ Test Abs.eval(). """
+        # Test Abs.eval().
         x = myokit.Abs(myokit.Number(0.9))
         self.assertEqual(x.eval(), 0.9)
         x = myokit.Abs(myokit.Number(-0.9))
         self.assertEqual(x.eval(), 0.9)
 
     def test_eval_unit(self):
-        """
-        Test Abs.eval_unit().
-        """
+        # Test Abs.eval_unit().
+
         # Mini model
         m = myokit.Model()
         c = m.add_component('c')
@@ -2005,26 +2098,24 @@ class AbsTest(unittest.TestCase):
             y.rhs().eval_unit(myokit.UNIT_STRICT), myokit.units.Newton)
 
     def test_tree_str(self):
-        """ Test Abs.tree_str(). """
+        # Test Abs.tree_str().
         x = myokit.Abs(myokit.Number(0.5))
         self.assertEqual(x.tree_str(), 'abs\n  0.5\n')
 
 
 class EqualTest(unittest.TestCase):
-    """
-    Tests myokit.Equal.
-    """
+    """ Tests myokit.Equal. """
+
     def test_eval(self):
-        """ Test Equal.eval(). """
+        # Test Equal.eval().
         x = myokit.Equal(myokit.Number(1), myokit.Number(1))
         self.assertTrue(x.eval())
         x = myokit.Equal(myokit.Number(1), myokit.Number(2))
         self.assertFalse(x.eval())
 
     def test_eval_unit(self):
-        """
-        Test Equal.eval_unit().
-        """
+        # Test Equal.eval_unit().
+
         # Mini model
         m = myokit.Model()
         c = m.add_component('c')
@@ -2067,7 +2158,7 @@ class EqualTest(unittest.TestCase):
         self.assertEqual(z.rhs().eval_unit(s), myokit.units.dimensionless)
 
     def test_tree_str(self):
-        """ Test Equal.tree_str(). """
+        # Test Equal.tree_str().
         x = myokit.Equal(myokit.Number(1), myokit.Number(2))
         self.assertEqual(x.tree_str(), '==\n  1\n  2\n')
         x = myokit.Plus(myokit.Number(3), x)
@@ -2075,9 +2166,7 @@ class EqualTest(unittest.TestCase):
 
 
 class NotEqualTest(unittest.TestCase):
-    """
-    Tests myokit.NotEqual.
-    """
+    """ Tests myokit.NotEqual. """
     def test_eval(self):
         # Test NotEqual.eval().
         x = myokit.NotEqual(myokit.Number(1), myokit.Number(1))
@@ -2094,9 +2183,7 @@ class NotEqualTest(unittest.TestCase):
 
 
 class MoreTest(unittest.TestCase):
-    """
-    Tests myokit.More.
-    """
+    """ Tests myokit.More. """
     def test_eval(self):
         # Test More.eval().
         x = myokit.More(myokit.Number(1), myokit.Number(1))
@@ -2113,9 +2200,7 @@ class MoreTest(unittest.TestCase):
 
 
 class LessTest(unittest.TestCase):
-    """
-    Tests myokit.Less.
-    """
+    """ Tests myokit.Less. """
     def test_eval(self):
         # Test Less.eval().
         x = myokit.Less(myokit.Number(1), myokit.Number(1))
@@ -2132,9 +2217,7 @@ class LessTest(unittest.TestCase):
 
 
 class MoreEqualTest(unittest.TestCase):
-    """
-    Tests myokit.MoreEqual.
-    """
+    """ Tests myokit.MoreEqual. """
     def test_eval(self):
         # Test MoreEqual.eval().
         x = myokit.MoreEqual(myokit.Number(1), myokit.Number(1))
@@ -2153,9 +2236,7 @@ class MoreEqualTest(unittest.TestCase):
 
 
 class LessEqualTest(unittest.TestCase):
-    """
-    Tests myokit.LessEqual.
-    """
+    """ Tests myokit.LessEqual. """
     def test_eval(self):
         # Test LessEqual.eval().
         x = myokit.LessEqual(myokit.Number(1), myokit.Number(1))
@@ -2174,9 +2255,8 @@ class LessEqualTest(unittest.TestCase):
 
 
 class AndTest(unittest.TestCase):
-    """
-    Tests myokit.And.
-    """
+    """ Tests myokit.And. """
+
     def test_eval(self):
         # Test And.eval().
         x = myokit.And(myokit.Number(1), myokit.Number(1))
@@ -2243,9 +2323,8 @@ class AndTest(unittest.TestCase):
 
 
 class OrTest(unittest.TestCase):
-    """
-    Tests myokit.Or.
-    """
+    """ Tests myokit.Or. """
+
     def test_eval(self):
         # Test Or.eval().
         x = myokit.Or(myokit.Number(1), myokit.Number(1))
@@ -2314,9 +2393,8 @@ class OrTest(unittest.TestCase):
 
 
 class NotTest(unittest.TestCase):
-    """
-    Tests myokit.Not.
-    """
+    """ Tests myokit.Not. """
+
     def test_code(self):
         # Test Not.code().
         x = myokit.Not(myokit.Number(1))
@@ -2372,9 +2450,8 @@ class NotTest(unittest.TestCase):
 
 
 class IfTest(unittest.TestCase):
-    """
-    Tests myokit.If.
-    """
+    """ Tests myokit.If. """
+
     def test_creation(self):
         # Test creation plus some accessor methods.
 
@@ -2466,9 +2543,8 @@ class IfTest(unittest.TestCase):
 
 
 class PiecewiseTest(unittest.TestCase):
-    """
-    Tests myokit.Piecewise.
-    """
+    """ Tests myokit.Piecewise. """
+
     def test_creation(self):
         # Test Piecewise creation plus some accessor methods.
 
@@ -2697,9 +2773,8 @@ class EquationTest(unittest.TestCase):
 
 
 class UserFunctionTest(unittest.TestCase):
-    """
-    Tests :class:`UserFunction`.
-    """
+    """ Tests :class:`UserFunction`. """
+
     def test_user_function(self):
         # Test :class:`UserFunction` creation and methods.
